[Test] Split unittest_htp_kernels into per-op files under htp/
Reorganize HTP unit tests from a single 1114-LoC file into a dedicated
test/unittest/htp/ subdirectory with per-op files:

  htp/unittest_htp_common.{h,cpp}  shared helpers (permute, ChanCtx,
                                   Q4_0 quant utils) and DSP session main()
  htp/unittest_htp_mat_mul.cpp     matmul variants: pwf16 RPC, pwf16 chan,
                                   wf16 RPC, pwqk0 RPC
  htp/unittest_htp_rms_norm.cpp    rms_norm_f32
  htp/unittest_htp_quantizer.cpp   repack_q4_0x4 -> x4x2
  htp/meson.build                  builds 3 test binaries, main meson.build
                                   invokes via subdir('htp')

Test cases are reduced from 97 to 35 by keeping representative dimensions
(small/medium/large square, GEMV, K>N and K<N rectangular, non-power-of-2 M)
that exercise each code path instead of enumerating every combination.

This follows the existing compiler/ and layers/ subdirectory pattern and
keeps test/unittest/ root uncluttered. Matching RPC and chan pwf16 cases
live in the same binary so MSE and round-trip latency can be compared
directly.
